revert debian basefiles to 13.2
#68 introduces a bump of the basefiles to 13.3. The changelog for this
version has a single entry:

base-files (13.3) unstable; urgency=medium

  [ Helmut Grohne ]
  * DEP17: Install /usr-merge aliasing symlinks. Closes: #1064459.

 -- Santiago Vila <sanvila@debian.org>  Thu, 06 Jun 2024 00:35:00 +0200

The link to the aforementioned issue with more details:
https://bugs-devel.debian.org/cgi-bin/bugreport.cgi?bug=1064459

In short it introduces symlinks in the root folder. Here is how the root
looks in 13.2:

bin/
boot/
dev/
etc/
home/
lib/
proc/
root/
run/
sbin/
sys/
tmp/
usr/
var/

And this is 13.3:
bin@
boot/
dev/
etc/
home/
lib@
proc/
root/
run/
sbin@
sys/
tmp/
usr/
var/


As seen, some toplevel folders are symbolic links. Now this is not a
problem for the regular debian root, but it is a problem for a base
container image. Since the Dockerfile COPY directive does not work well
with folder symlinks.

Therefore here we revert it back to 13.2, and we probably need to pin it like this.


Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@isovalent.com>